refactor: move function from oneclick mall data file to deferred
